+++ src/Query/Query.php
@@ -285,6 +285,21 @@ class Query {
     }
 
     /**
+     * Get a message by its message number
+     * @param $msgn
+     * @param null $msglist
+     * 
+     * @return Message
+     * @throws ConnectionFailedException
+     * @throws InvalidMessageDateException
+     * @throws MessageContentFetchingException
+     * @throws MessageHeaderFetchingException
+     * @throws RuntimeException
+     * @throws \Webklex\PHPIMAP\Exceptions\EventNotFoundException
+     */
+    public function getMessageByMsgn($msgn, $msglist = null){
+        return $this->getMessage($msgn, $msglist, IMAP::ST_MSGN);
+    }
      * Paginate the current query
      * @param int $per_page
      * @param null $page